Updated nextavailableshare default for the pph object
The default value for the nextavailableshare property of the pph object
used to be none. This yields errors in the python33 interpreter (at
least) when doing a max call, since the comparison is made between a
None type and an Integer type.

The new default is 1. In case the store was already saved, it will be
updated from the values in the file.