Added support for python 3 byte-arrays
Python 2 used to support byte arrays as an identical equilvalent of
strings. We could append and manage bytes directly into the string
objects transparently throughout the sourcecode. This is not the case in
python 3.

I added the int.to_byte array method for the bytes in the secret. This
is intended to force data types into our the thresholdless key and the
shamir secret data.

At the time of this commit, we are not forcing checks on secret_data or
thresholdless key input. WE SHOULD CHECK THE INPUT on this matter so
strings are not given to us. However, even if the check should be made.
The secrets that we are generating on the pph layer are byte arrays by
definition.

All tests pass on py27 and py33.